CHANGES:
Rewrite module. Arranged into a function. Previously, it unloaded all the values from the .txt file and created the dictionary. It was stored in RAM. Now it processes the request and returns the value of the key, only if it equally to the value from the request.

Corrected the description of the module.

ADDS:

Added a function call if it runs directly from the module.

Added the description to the function.